+++ nestedtext.py
@@ -36,8 +36,8 @@ from inform import (
 
 
 # Globals {{{1
-__version__ = "0.3.1"
-__released__ = "2020-09-05"
+__version__ = "0.3.2"
+__released__ = "2020-09-07"
 __all__ = ['loads', 'dumps', 'NestedTextError']